fix ex00/ex01 and ex02 done

+++ cpp09/ex00/BitcoinExchange.hpp
@@ -7,13 +7,13 @@
 
 class BitcoinExchange {
     private:
-        std::map<std::string, float> input_csv;
+        std::map<std::string, double> input_csv;
     public:
         BitcoinExchange();
         ~BitcoinExchange();
         BitcoinExchange(const BitcoinExchange &btc);
         BitcoinExchange &operator=(const BitcoinExchange &btc);
         void processingData(std::string file);
-        float getDateValue(std::string date);
+        double getDateValue(std::string date);
 };
 # endif
